// src/web/lib/api.ts
export interface ICreatorShareDto {
    player: PlayerPartialDto;
    quantity: number;
}
// ...
export enum Platform {
    Twitch = "Twitch",
}

/** Mirrors `TTX.App.Dto.Players.PlayerPartialDto`. */
export class PlayerPartialDto implements IPlayerPartialDto {
    avatar_url!: string;
    created_at!: Date;
    credits!: number;
    id!: number;
    name!: string;
    platform!: Platform;
    platform_id!: string;
    platform_url!: string;
    portfolio!: number;
    slug!: string;
    type!: PlayerType;
    updated_at!: Date;
    value!: number;

    constructor(data?: IPlayerPartialDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.avatar_url = _data["avatar_url"] !== undefined ? _data["avatar_url"] : null as any;
            this.created_at = _data["created_at"] ? new Date(_data["created_at"].toString()) : null as any;
            this.credits = _data["credits"] !== undefined ? _data["credits"] : null as any;
            this.id = _data["id"] !== undefined ? _data["id"] : null as any;
            this.name = _data["name"] !== undefined ? _data["name"] : null as any;
            this.platform = _data["platform"] !== undefined ? _data["platform"] : null as any;
            this.platform_id = _data["platform_id"] !== undefined ? _data["platform_id"] : null as any;
            this.platform_url = _data["platform_url"] !== undefined ? _data["platform_url"] : null as any;
            this.portfolio = _data["portfolio"] !== undefined ? _data["portfolio"] : null as any;
            this.slug = _data["slug"] !== undefined ? _data["slug"] : null as any;
            this.type = _data["type"] !== undefined ? _data["type"] : null as any;
            this.updated_at = _data["updated_at"] ? new Date(_data["updated_at"].toString()) : null as any;
            this.value = _data["value"] !== undefined ? _data["value"] : null as any;
        }
    }

    static fromJS(data: any): PlayerPartialDto {
        data = typeof data === 'object' ? data : {};
        let result = new PlayerPartialDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["avatar_url"] = this.avatar_url !== undefined ? this.avatar_url : null as any;
        data["created_at"] = this.created_at ? this.created_at.toISOString() : null as any;
        data["credits"] = this.credits !== undefined ? this.credits : null as any;
        data["id"] = this.id !== undefined ? this.id : null as any;
        data["name"] = this.name !== undefined ? this.name : null as any;
        data["platform"] = this.platform !== undefined ? this.platform : null as any;
        data["platform_id"] = this.platform_id !== undefined ? this.platform_id : null as any;
        data["platform_url"] = this.platform_url !== undefined ? this.platform_url : null as any;
        data["portfolio"] = this.portfolio !== undefined ? this.portfolio : null as any;
        data["slug"] = this.slug !== undefined ? this.slug : null as any;
        data["type"] = this.type !== undefined ? this.type : null as any;
        data["updated_at"] = this.updated_at ? this.updated_at.toISOString() : null as any;
        data["value"] = this.value !== undefined ? this.value : null as any;
        return data;
    }
}

/** Mirrors `TTX.App.Dto.Players.PlayerPartialDto`. */
export interface IPlayerPartialDto {
    avatar_url: string;
    created_at: Date;
    credits: number;
    id: number;
    name: string;
    platform: Platform;
    platform_id: string;
    platform_url: string;
    portfolio: number;
    slug: string;
    type: PlayerType;
    updated_at: Date;
    value: number;
}
// ...
export enum PlayerType {
    User = "User",
    Admin = "Admin",
}